Implements onboarding flow
Adds an onboarding flow for new users, which includes profile creation and circle association.

Adds a 'notOnboarded' state to the authentication status to direct new users to the onboarding screens.

Renames and restructures the login view and route to support circle onboarding.

This change allows users to either create a new circle or join an existing one using an invite code, improving the user experience for new users.